pyproject.toml
[tool.poetry]
name = "trading-experiment"
version = "0.1.0"
description = ""
authors = ["Levan Begashvili <life.long.learner127@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.22"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
ta-lib = "^0.5.1"
yfinance = "^0.2.49"
scipy = "^1.14.1"
torch = "^2.5.1"
scikit-learn = "^1.5.2"
tensorflow = "^2.18.0"
xgboost = "^2.1.3"
plotly = "^5.24.1"
streamlit = "^1.40.2"
psycopg2 = "^2.9.10"


[tool.poetry.group.dev.dependencies]
flake9 = "^3.8.3.post2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,"
exclude = "*migrations*"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
